refactor(gate): split the 1049-line detect-gate.ts god file (review item 1)
Pure move, no logic change — behaviour-preserving (all existing gate/loop/session
tests pass unchanged). detect-gate.ts split into focused src/gate/ modules:
  types.ts        IGate, IFileLintProblem, FileLinter
  tool-paths.ts   resolveToolBin + the bundled BIN/CONFIG/CHECK paths
  tsconfig.ts     strict tsconfig overlays + ensureWebGateTsconfig + tscPart
  shell.ts        shSingleQuote + packEnvPrefix
  test-discovery.ts  discoverTestCommand + webTestProbe
  linter.ts       makeFileLinter + formatFile + prettierWriteCommand
  core-gate.ts    buildGate + buildCoreFix
  web-gate.ts     buildWebGate/TypeGate/TscCheck/Fix + WEB_FRAMEWORKS/WEB_PACKS
  index.ts        public barrel
Web scaffolding (scaffoldWeb/installWebDeps/webGuidance/BUILD_PREAMBLE) → src/scaffold/web-scaffold.ts.

Path constants moved a directory deeper, so import.meta.dir joins gained a level
(verified STRICT_CONFIG/STRICT_WEB_CONFIG/BROWSER_CHECK resolve to the package
root). Updated all 18 import sites (4 src, 3 scripts, 11 tests).
