+++ packages/core/src/gate/types.ts
@@ -1,0 +1,29 @@
+/**
+ * Build the gate that confirms "done" — and makes tsforge a TypeScript-SPECIALIZED
+ * harness, not a generic file editor. It enforces strict TS on whatever the model
+ * writes, in two layers, using tsforge's OWN bundled toolchain so it works on any
+ * target regardless of that project's setup:
+ *   1. `tsc --strict --noUncheckedIndexedAccess` — the TYPE-aware floor (unguarded
+ *      `arr[i]`, null-safety, real type errors). Greenfield gets a strict tsconfig
+ *      brought in; an existing project's own tsconfig is respected.
+ *   2. the bundled eslint strict config — the SYNTACTIC idioms (no `as`/`any`/`!`,
+ *      no over-annotation), which need no type info or deps.
+ * The deterministic gate loop + rule-docs cards + ast-grep polish then drive the
+ * local model's output up to that bar — that's the uplift.
+ */
+export interface IGate {
+  /** The shell command run to verify (must exit 0). */
+  command: string;
+  /** A short human label for the banner. */
+  label: string;
+}
+
+/** One lint violation on a single file (errors only), for write-time feedback. */
+export interface IFileLintProblem {
+  line: number;
+  message: string;
+  ruleId: string;
+}
+
+/** Lint ONE just-written file, returning its errors. Reused per write. */
+export type FileLinter = (absPath: string) => Promise<IFileLintProblem[]>;
