+++ packages/core/tests/web-gate-tsconfig.test.ts
@@ -2,7 +2,7 @@ import { test, expect } from "bun:test";
 import { mkdtemp, rm, mkdir, writeFile } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
-import { buildWebGate, buildWebTypeGate } from "../src/detect-gate";
+import { buildWebGate, buildWebTypeGate } from "../src/gate";
 
 // Issue: a `bun:test` import in a scaffolded web app reds the gate with TS2307
 // ("Cannot find module 'bun:test'"). Root cause: the web gate ran `tsc -p
